+++ ms-pagos/src/main/java/estacionamientos/ms_pagos/dto/BancoDTO.java
@@ -1,5 +1,7 @@
 package estacionamientos.ms_pagos.dto;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
@@ -15,5 +17,5 @@ public class BancoDTO {
     private String nombre;
 
     @NotNull
-    private Double descuento;
+    private BigDecimal descuento;
 }
