+++ ms-pagos/src/main/java/estacionamientos/ms_pagos/dto/CobroResponseDTO.java
@@ -1,5 +1,6 @@
 package estacionamientos.ms_pagos.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
@@ -16,8 +17,8 @@ public class CobroResponseDTO {
     private Long id;
     private Long idAcceso;
     private Long idCliente;
-    private Long minutos;
-    private Double montoBase;
+    private Integer minutos;
+    private Float montoBase;
     private Double montoFinal;
     @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
     private LocalDateTime fechaCobro;
